Add files via upload
Car Rental Website Project
Introduction
This is a car rental website project.
The purpose of this project is to create a car rental website.
The website will have a login page, a car listing page, a car reservation page, and a car return page.
The login page will have a username and password input field.
The car listing page will have a search bar and a car listing table.
The car reservation page will have a car reservation table.
Requirements
XAMPP is required to run the project.
The project will be hosted on a localhost server.
Installation
Install XAMPP.
Open XAMPP and click on the Start button.
Clone the project to the root of the XAMPP server.
Open the project in XAMPP.
Import SQL file from database folder to the XAMPP server.
Open Browser and navigate to localhost:8080.
Hurray! The project is now running.
Database
The database is stored in a folder called database.
The database is named as carproject.sql.
The database is stored in the root of the XAMPP server.
The database is imported to the XAMPP server.
Database used is MySQL.
Database Connection page is named connection.php.